Add basic panning and zoom
It seems to work, but it is not pleasant to use.
Booth the panning and scrolling speed should scale with the zoom factor.
At the moment the more we are zoomed in, the more insane the panning
speed gets, and slower scrolling is.
It should be constant all the way.

Additionally we can zoom ourselves into the negative range in which case
the whole screen flips, which is rather hilarious.