Fix the view not scaling properly when changing monitors
Likely a noob mistake, but it was a rather big head scratcher for
me so I have written down the observed issue and the method how I
have figured out what was wrong.

The issue was that when the window was initially created the view
into the mandelbrot set would be scaled properly. This would be
the case regardless on which monitor the window was initialy created on.
But the views scaling would get messed up if the window would get
dragged onto the other monitor.
The issue was that while winit knew about the new necessary resolution
the GPU didn't. I didn't know that the surface should be recreated with
the new resolution as that seems to determine the size of the texture
the GPU is rendering to.