Zoom and transition deltas now scale with the level of zoom
For increasing zoom level we expect the panning speed to slow down,
while the zooming speed to increase.
The reason is simple. The closer we are, the less we need to move
to experience large movements. Similarly, the closer we are, more
we need to zoom for the same level of perceived zoom increase.