Add scaffolding for switching between CPU/GPU rendering
The user may switch between the two by pressing the 'c' and 'g'
keys on their keyboard.
The reaction to the keypress only occurs when the key is released.
This is less responsive, but we resolve a number of problems.
While the Pressed event triggers as long the key is pressed the
Release event only triggers once.
Eliminating the need for safeguarding against multiple consecutive
triggers.