Add basic CPU based rendering
This solution uses a the softbuffer library to draw to the raw window.
The user may switch between CPU and GPU based rendering, but there seems
to be something in the GPU based implementation which will keep holding
and refreshing the last GPU image even if the user switches to CPU based
rendering. Not sure what may cause this. Maybe the whole render pipeline
would have to be destroyed when switching to the CPU?

As a start if the CPU rendering is selected it will work.
As of yet CPU based rendering does not support zooming or panning.