Add basic chunking support for CPU rendering
It was simple to write, but the funny thing is, is that it isn't
much faster than the single threaded version.
In release mode a whole image is rendered in about a half a second,
while with this chunking perhaps in about 35 ms.
None of this was measured, just a feeling.

Probebbly a worker pool would be better suited for the task, as it
would not have to start and clean up the threads after each rendering
cycle.