Prototype rewrite of GPU based rendering
Now the GPU is not used directly to draw to the screen, but instead
it renders into a texture which we then display the same way on the
screen as we would during CPU based rendering.
This resolves the issue not being able to switch between the two
rendering modes. Now that works properly.

It is a little nasty and the GPU part is a bit too slow to my liking.
Not sure where the delays are coming from. Sure we are copying the
resulting texture at least twice, but thay shouldn't be that big of
an issue.