Add synchronization for the map_async callback
According to the wgpu documentation we know that device.poll will be
executing the callback function supplied to map_async.
We know that we are on one thread (for this section) so there is no
real for syncronization here.
Non the less it is inserted in case someone accidentally finds this
piece of code and based on it forgets to add it.