Clean up softbuffer resizing on WindowEvent::Resized
Techinacally we could modify the code so that it proceeds to chug
along even if the buffer for the softbuffer could not have been resized.
However this introduces more complexity and is likely pointless.
In the case where we don't have enough memory to resize this buffer,
likely we will have other memory problems in later steps as well.