Fix synchronization added in commit 9c10bad
In that commit an atomic boolean value was added in haste to handle
synchronization.
That implementation doesn't synchronize at all. Not sure what I was
doing.

Now we have a proper mutex, condvar combination. The calling thread
will sleep as long as it needs to until it is woken up by the result,
which will then be available for processing.
This returns the proper value and the error value as well while
efficiently waiting for the result.
Technically it would be prudent to wait with a timeout, but
at some point we must accept that things occassionally break.

What I wasn't expecting that this implementation would actually
run faster.
The atomic or the one prior would take a max of 15 ms to render a
frame using the GPU. With this synchronization in place, somehow
the highest frame time I have observed was around 8 ms.
Interesting.